+++ src/main.js
@@ -1,0 +1,3 @@
+import vcCheckbox from './Checkbox'
+
+module.exports = vcCheckbox