+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 <h1 align="center">vc-checkbox - @changelog</h1>
 
+## 0.2.0
+
+`2016-09-18`
+
+- add `inline` prop and default to `true`
+  - 不设置或者设置为任意不为`false`的值*并且*处在buttonGroup模式时，自动内联
+  - 用于实现紧密并列的buttonGroup效果
+
 ## 0.1.0
 
 `2016-09-01`